pyproject.toml
[project]
name = "collabia"
version = "0.1.0"
description = "Multi-agent conversational chatbot where LLMs debate until reaching consensus"
requires-python = ">=3.11"
dependencies = [
    "anthropic[vertex]>=0.49.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.60.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
collabia = "collabia.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
